+++ src/core/executables.js
@@ -19,12 +19,13 @@ const CP = "/bin/cp";
 const LN = "/bin/ln";
 const MV = "/bin/mv";
 const RM = "/bin/rm";
+const SH = "/bin/sh";
 const TEST = "/bin/test";
 const MKTEMP = "/usr/bin/mktemp";
 const PRINTENV = "/usr/bin/printenv";
 const STAT = "/usr/bin/stat";
 
-const EXECUTABLES = [CAT, CP, LN, MV, RM, TEST, MKTEMP, PRINTENV, STAT];
+const EXECUTABLES = [CAT, CP, LN, MV, RM, SH, TEST, MKTEMP, PRINTENV, STAT];
 
 module.exports = {
     CAT,
@@ -32,6 +33,7 @@ module.exports = {
     LN,
     MV,
     RM,
+    SH,
     TEST,
     MKTEMP,
     PRINTENV,